Add diverging scale for #65
We need to be able to set diverging scales by parameter so we can add a "diverging: true" property to the parameter config.json file. The colour for the negative side of a diverging scale needs to be set per scenario using a "negativecolor" property in the scenarios index.json file.